+++ packages/core/CHANGELOG.md
@@ -1,5 +1,12 @@
 # internal-platform
 
+## 2.0.11
+
+### Patch Changes
+
+- 302bd02f: Issue #377: only expose the external eventId in the API
+- b5db9f5e: Adding MIT license
+
 ## 2.0.10
 
 ### Patch Changes